requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-import-export==2.8.0
et-xmlfile==1.1.0
frozenlist==1.3.0
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
gspread==5.4.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
line-bot-sdk==2.3.0
MarkupPy==1.14
multidict==6.0.2
numpy==1.23.0
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.4.3
protobuf==4.21.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pygsheets==2.0.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
yarl==1.7.2
zipp==3.8.0
